requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Network & Security
scapy==2.5.0
python-nmap==0.7.1
netifaces==0.11.0
pyshark==0.6

# AI & Machine Learning
openai==1.6.1
google-generativeai>=0.8.0
langchain==0.1.0
langchain-openai==0.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Database & Caching
motor==3.3.2  # Async MongoDB
redis==5.0.1
pymongo==4.6.1

# WebSocket & Real-time
python-socketio==5.10.0
aiohttp==3.9.1
websockets==12.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
orjson==3.9.10
httpx==0.25.2

# Logging & Monitoring
loguru==0.7.2
rich==13.7.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
